Another way to open Sidecar
When exiting Sidecar in Control Center. The checkbox is not selected. But Sidecar is actually still connected. This is a probability. I guess it is a system bug.
Here is another way to open Sidecar. Go to display settings in system settings to start or exit Sidecar. It is a simple code and tested on Monterey without problems.